+++ HISTORY.md
@@ -85,6 +85,7 @@
 
 ### 2023-10-12
 * updated inline.html test page with "noDevWarning" configuration
+* objects: tweaked sprite name-change user edit event details
 
 ### 2023-10-11
 * objects: added serialized scripts (w/o dependencies) to scripts-related user-logging events